Move image format detection to new file
This file will grow as we add more sophisticated format detection
